+++ src/core/classes/finish_conditions/__init__.py
@@ -1,3 +1,6 @@
 from .final_states import *
 from .repetition import Repetition
 from .check import Check
+from .dead_position import DeadPosition
+from .finish_condition import FinishCondition
+from .fifty_moves import FiftyMoves
