+++ src/core/classes/game_modes/standard.py
@@ -1,4 +1,4 @@
-from ..finish_conditions.fifty_moves import FinishCondition, FiftyMoves
+from ..error_classes import InvalidMove
 from ..action_rules import *
 from ..board import *
 from ..pieces import *
@@ -6,13 +6,25 @@ from ..rules import *
 from ..action_rules import *
 from ..color import *
 from ..finish_conditions import *
+from ..ply import Ply
 from .game_mode import GameMode
 
 
 class Standard(GameMode):
     rules : list[Rule] = [ValidColor(), ValidPieceMove(), TakeOwnPiece(), Block(), Pin()]
     action_rules : list[ActionRule] = [PromoteRule(), EnPassantRule(), CastleRule(), MoveRule()]
-    finish_conditions : list[FinishCondition] = [Repetition(), FiftyMoves(), Check()]
+    finish_conditions : list[FinishCondition] = [Repetition(), FiftyMoves(), DeadPosition()]
+
+    def __init__(self):
+        self.connect_rules()
+
+    def connect_rules(self) -> None:
+        for list_of_rules in [self.rules, self.action_rules, self.finish_conditions]:
+            previous_rule = None
+            for rule in list_of_rules:
+                if previous_rule:
+                    previous_rule.set_next(rule)
+                previous_rule = rule
 
     def init_board(self) -> Board:
         board = Board()
@@ -22,51 +34,27 @@ class Standard(GameMode):
                         board.add(piece_class(color), Position(col,row))
         return board
 
-    # def validate(self, ply: Ply, board: Board) -> bool:
-    #     """
-    #     Validates a ply for the current board and a previous board. Returns `True` if the ply is valid
-    #     and `False` if it's not valid. If the ply is valid, this method fills the action attribute for
-    #     the ply.
-    #     """
-    #     piece = board.get_piece(ply.from_position)
+    def validate(self, ply: Ply, board: Board) -> bool:
+        try:
+            self.rules[0].handle(ply, board)
+            self.action_rules[0].handle(ply, board)
+        except InvalidMove:
+            return False
+        except Exception as err:
+            raise err
+        else:
+            return True
 
-    #     if not piece.can_move(ply.vector):
-    #         return False
-        
-    #     ply.action = self.get_action(ply, board)
-
-    #     current_board = board.copy()
-    #     piece_taken = current_board.get_piece(ply.to_position)
-
-    #     # Exclude taking your own piece, pins and blocks
-    #     if piece_taken and piece_taken.color == ply.color or \
-    #        not isinstance(piece, Knight) and self.is_blocked(ply, current_board) or \
-    #        self.is_pinned(ply, board):
-    #         return False
-
-    #     # Pawns can only move diagonally under certain conditions
-    #     elif isinstance(piece, Pawn):
-    #         if piece_taken:
-    #             if (isinstance(ply.action, Move) or isinstance(ply.action, Promote)) and \
-    #                 piece.is_capture(ply.to_position - ply.from_position):
-    #                 return True
-    #             else:
-    #                 return False
-    #         elif isinstance(ply.action, EnPassant) or not piece.is_capture(ply.to_position - ply.from_position):
-    #             return True
-    #         else:
-    #             return False           
-
-    #     # Can't castle if there is a check in the way
-    #     if isinstance(ply.action, Castle):
-    #         pointer = ply.color.king_position
-    #         direction = ply.action.direction
-    #         for n in range(0,2):
-    #             pointer += direction
-    #             possible_board = board.try_action(Move(piece, direction))
-    #             if self.is_check(ply.color, possible_board):
-    #                 return False
-    #     return True
-
-    def is_win(self, color: Color, board: Board):
-        return Check.is_check(color.opposite_color(), board)
+    def is_finished(self, ply: Ply, board: Board) -> FinalState:
+            final_state = self.finish_conditions[0].handle(ply, board)
+            if final_state:
+                return final_state
+            else:
+                for piece in board.get_pieces(color=ply.color.opposite_color(), is_active=True):
+                    if len(self.get_valid_moves(piece, board)) > 0:
+                        return None
+                else:
+                    if Check.is_check(ply.color.opposite_color(), board):
+                        return Win()
+                    else:
+                        return Draw()
